requirements.txt
langchain>=0.0.300  # Updated, and using >= for minor version updates
langchain-community>=0.0.300 # Updated
streamlit>=1.22.0 # Updated
tiktoken>=0.4.0 # Updated
faiss-cpu>=1.7.4 
openai>=0.28.0 # Corrected capitalization and using >=.
protobuf~=3.19.0
mysql-connector-python
pymysql
sentence-transformers
chromadb
langchain_google_genai
python-dotenv~=1.0.1
urllib3~=2.3.0